Handwritten_Character_Recognition/requirements.txt

tensorflow
numpy
opencv-python
matplotlib
streamlit
pillow
